pyproject.toml
[project]
name = "fly-ai"
version = "0.1.0"
description = "FLY.AI B3 Data Source with DDD and Clean Architecture"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "playwright>=1.40.0",
    "fastapi>=0.100.0",
    "psycopg2-binary>=2.9.11",
    "uvicorn[standard]>=0.41.0",
    "pandas>=3.0.1",
    "numpy>=2.4.3",
    "pydantic-settings>=2.13.1",
    "unidecode>=1.4.0",
    "prometheus-client>=0.24.1",
    "pact-python>=3.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-playwright>=0.4.0",
    "mutmut>=2.4.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shared", "src/companies", "src/financials", "src/market_data"]
# SOTA: Define the source directory for the packages
[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
pythonpath = [
    "/home/stangler/Documents/Python/FLY.AI/src", "."
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mutmut>=3.5.0",
    "pytest-asyncio>=1.3.0",
]


[tool.mutmut]
paths_to_mutate = ["src/companies/domain/entities/", "src/shared/infrastructure/utils/"]
tests_dir = ["tests/companies/domain/entities/", "tests/shared/infrastructure/utils/"]
debug = true
# SOTA Rule: Use relative PYTHONPATH so mutmut uses the mutated files in the sandbox
runner = "PYTHONPATH=src uv run pytest -x"
